Changed env, wrappers, network architectures, storage, to support convolutional networks. Results for env comparison included and trained models. Also saving percentage of boards routed to enable curriculum learning.

+++ environment/PCBBoard.py
@@ -57,7 +57,7 @@ class PCBBoard(abc.ABC, gym.Env):
         self.MAX_OBS_COLS = 10
         self.middle_obs_row = self.MAX_OBS_ROWS // 2
         self.middle_obs_col = self.MAX_OBS_COLS // 2
-        self.grid = np.full(shape=(grid_rows, grid_cols), fill_value=obstacle_value, dtype=np.float32)
+        self.grid = np.full(shape=(grid_rows, grid_cols), fill_value=obstacle_value, dtype=np.int)
         self.rows = rows
         self.cols = cols
         self.obs_rows = self.MAX_OBS_ROWS
